refactor: 사용자가 3자리의 숫자를 입력하는 기능 수정
